+++ packages/api/challenges/feedback/speech.js
@@ -121,20 +121,7 @@ export function feedback(
 ) {
   return prepare(challengeId)
     .then(feedbackId =>
-      listenAndReply(
-        feedbackId,
-        progressiveResultsCb,
-        recorder,
-        dataEvent,
-      ).then(result => {
-        // Close the connection before returning the result;
-        closeWebsocketConnection();
-        return result;
-      }),
+      listenAndReply(feedbackId, progressiveResultsCb, recorder, dataEvent),
     )
-    .catch(error => {
-      // Try to close the connection before returning the error;
-      closeWebsocketConnection();
-      return error;
-    });
+    .finally(() => closeWebsocketConnection());
 }